Improve structure of utilities.backend
